Fix error processing user inactive message.
A recent changed made is passing the context to the SDK when processing
the input so it can be reused.

When a user inactive message is processed,
a new input message is created with an empty text and this new message
should be processed instead.

This changes fixes this by setting the context passed as parameter to
null when a user inactive message arrives so a new context will be
created in this case with the new message.